sos_service/requirements.txt
python-dotenv
typing
requests

Flask
Flask-Cors
Werkzeug

confluent-kafka
kafka-python
kafka-python-ng

supabase
pymongo
pydantic
